Update to address PR feedback
