+++ php/src/Google/Protobuf/Internal/GPBUtil.php
@@ -32,11 +32,20 @@
 
 namespace Google\Protobuf\Internal;
 
+use Google\Protobuf\BoolValue;
+use Google\Protobuf\BytesValue;
+use Google\Protobuf\DoubleValue;
 use Google\Protobuf\Duration;
 use Google\Protobuf\FieldMask;
+use Google\Protobuf\FloatValue;
+use Google\Protobuf\Int32Value;
+use Google\Protobuf\Int64Value;
 use Google\Protobuf\Internal\GPBType;
 use Google\Protobuf\Internal\RepeatedField;
 use Google\Protobuf\Internal\MapField;
+use Google\Protobuf\StringValue;
+use Google\Protobuf\UInt32Value;
+use Google\Protobuf\UInt64Value;
 
 function camel2underscore($input) {
     preg_match_all(
@@ -193,6 +202,9 @@ class GPBUtil
 
     public static function checkMessage(&$var, $klass, $newClass = null)
     {
+        if (self::isWrapperType($klass)) {
+            self::normalizeToMessageType($var, $klass);
+        }
         if (!$var instanceof $klass && !is_null($var)) {
             throw new \Exception("Expect $klass.");
         }
@@ -596,5 +608,67 @@ class GPBUtil
                is_a($msg, "Google\Protobuf\BoolValue")   ||
                is_a($msg, "Google\Protobuf\StringValue") ||
                is_a($msg, "Google\Protobuf\BytesValue");
+    }
+
+
+    private static $wrapperTypes = [
+        DoubleValue::class,
+        FloatValue::class,
+        Int64Value::class,
+        UInt64Value::class,
+        Int32Value::class,
+        UInt32Value::class,
+        BoolValue::class,
+        StringValue::class,
+        BytesValue::class,
+    ];
+
+
+    /**
+     * Determine if a fully qualified class name is a wrapper type.
+     *
+     * @param string $class the fully qualified class name
+     * @return bool true if this is a wrapper type
+     */
+    public static function isWrapperType($class)
+    {
+        return in_array($class, self::$wrapperTypes);
+    }
+
+    /**
+     * Tries to normalize $value into a provided protobuf wrapper type $class.
+     * If $value is any type other than an object, we attempt to construct an
+     * instance of $class and assign $value to it using the setValue method
+     * shared by all wrapper types.
+     *
+     * @param mixed $value The value passed to the protobuf setter method
+     * @param string $class The expected wrapper class name
+     * @throws \Exception If $value cannot be converted to a wrapper type
+     */
+    public static function normalizeToMessageType(&$value, $class)
+    {
+        if (is_null($value) || is_object($value)) {
+            // This handles the case that $value is an instance of $class. We
+            // choose not to do any more strict checking here, relying on the
+            // existing type checking done by GPBUtil.
+            return;
+        } else {
+            // Try to instantiate $class and set the value
+            if (!self::isWrapperType($class)) {
+                throw new \Exception("Expected wrapper type, got '$class'");
+            }
+            try {
+                $msg = new $class;
+                $msg->setValue($value);
+                $value = $msg;
+                return;
+            } catch (\Exception $exception) {
+                throw new \Exception(
+                    "Error normalizing value to type '$class': " . $exception->getMessage(),
+                    $exception->getCode(),
+                    $exception
+                );
+            }
+        }
     }
 }
